Fix the backtracking algorithm for searching path in a grid
There was a small issue that I fixed. Something else what I did was to add
memoization for the implementation so that the alg. complexity is now
O(rows*cols) instead of O(2^(rows+cols)). I also refactored the code to be
a bit more OOP like.

Tested manually by running the alg. with different inputs.